incorporate training scripts into jupyter notebook
