+++ transformed_datasets/train/sentences/state.json
@@ -4,7 +4,7 @@
       "filename": "data-00000-of-00001.arrow"
     }
   ],
-  "_fingerprint": "1436b243eeb8fabe",
+  "_fingerprint": "fb67153c37874e5b",
   "_format_columns": [
     "FileName",
     "Sentence",
